Documents/soc-analyst-api-feature-api_real/soc-analyst-api/requirements.txt
click==8.1.7
Flask==2.2.5
itsdangerous==2.1.2
watchdog>=1.0.2
geopy>=2.4.1
PyJWT>=2.8.0
flask_cors
Flask-SQLAlchemy

#For AI prompt#
langchain
langchain-experimental
streamlit
langchain-groq
psycopg2-binary
langchain
openai
langchain-openai
meraki
pandasai
plotly
kaleido
PILLOW

# python3 -m venv myenv
# source myenv/bin/activate